crypto/kyber: enforce strict decode and split AEAD module
