+++ dsm_client/deterministic_state_machine/dsm/src/crypto/kyber.rs
@@ -1,4 +1,4 @@
-//! ML-KEM-768 (Kyber) post-quantum key encapsulation with AES-256-GCM helpers.
+//! ML-KEM-768 (Kyber) post-quantum key encapsulation.
 //!
 //! Provides the DSM protocol's key exchange mechanism using NIST-standardized
 //! ML-KEM-768 (formerly CRYSTALS-Kyber). This module is strictly bytes-only
@@ -10,23 +10,16 @@
 //! - [`generate_kyber_keypair`] -- generate an ML-KEM-768 keypair
 //! - [`kyber_encapsulate`] -- encapsulate a shared secret to a public key
 //! - [`kyber_decapsulate`] -- decapsulate a shared secret using a secret key
-//! - [`aes_encrypt`] / [`aes_decrypt`] -- AES-256-GCM authenticated encryption
-//!   using shared secrets derived from ML-KEM key exchange
 //!
 //! # Dependencies
 //!
 //! - `ml-kem = "0.2"` (ML-KEM-768 implementation)
-//! - `aes-gcm = "0.10"` (AES-256-GCM authenticated encryption)
 //! - `blake3` (domain-separated key derivation)
 //! - `zeroize` (secret key scrubbing)
 
 use crate::types::error::DsmError;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 
-use aes_gcm::{
-    aead::{Aead, KeyInit},
-    Aes256Gcm, Nonce,
-};
 use crate::crypto::blake3::dsm_domain_hasher;
 use ml_kem::{
     kem::{Decapsulate, DecapsulationKey, Encapsulate, EncapsulationKey},
@@ -106,7 +99,16 @@ impl KyberKeyPair {
 
         let (pk, rest) = decode_bytes(bytes)?;
         bytes = rest;
-        let (sk, _rest) = decode_bytes(bytes)?;
+        let (sk, rest) = decode_bytes(bytes)?;
+
+        if !rest.is_empty() {
+            return Err(DsmError::serialization_error(
+                "Trailing bytes in KyberKeyPair encoding",
+                "keypair_bytes",
+                None::<&str>,
+                None::<std::io::Error>,
+            ));
+        }
 
         if pk.len() != PK_LEN || sk.len() != SK_LEN {
             return Err(DsmError::crypto(
@@ -278,9 +280,9 @@ pub fn generate_kyber_keypair_from_entropy(
     entropy: &[u8],
     context: &str,
 ) -> Result<(Vec<u8>, Vec<u8>), DsmError> {
-    if entropy.len() < 16 {
+    if entropy.len() < 32 {
         return Err(DsmError::crypto(
-            "Insufficient entropy for secure key generation (minimum 16 bytes required)",
+            "Insufficient entropy for secure key generation (minimum 32 bytes required)",
             None::<std::io::Error>,
         ));
     }
@@ -454,70 +456,6 @@ pub fn kyber_decapsulate(secret_key: &[u8], ciphertext: &[u8]) -> Result<Vec<u8>
     Ok(shared_secret.as_slice().to_vec())
 }
 
-// ------------------ AES-GCM (no GenericArray::from_slice anywhere) ------------------
-
-pub fn aes_encrypt(key: &[u8], nonce: &[u8], plaintext: &[u8]) -> Result<Vec<u8>, DsmError> {
-    if key.len() != 32 {
-        return Err(DsmError::crypto(
-            "Invalid key size for AES-256".to_string(),
-            None::<std::io::Error>,
-        ));
-    }
-    if nonce.len() != 12 {
-        return Err(DsmError::crypto(
-            "Invalid nonce size for AES-GCM".to_string(),
-            None::<std::io::Error>,
-        ));
-    }
-
-    // Build cipher without touching GenericArray::from_slice
-    let cipher = Aes256Gcm::new_from_slice(key).map_err(|_| {
-        DsmError::crypto(
-            "Invalid AES-256 key length".to_string(),
-            None::<std::io::Error>,
-        )
-    })?;
-
-    let mut nonce_arr = [0u8; 12];
-    nonce_arr.copy_from_slice(nonce);
-    let nonce = Nonce::from(nonce_arr); // From<[u8;12]>, no removed APIs
-
-    cipher
-        .encrypt(&nonce, plaintext)
-        .map_err(|_| DsmError::crypto("AES encryption failed".to_string(), None::<std::io::Error>))
-}
-
-pub fn aes_decrypt(key: &[u8], nonce: &[u8], ciphertext: &[u8]) -> Result<Vec<u8>, DsmError> {
-    if key.len() != 32 {
-        return Err(DsmError::crypto(
-            "Invalid key size for AES-256".to_string(),
-            None::<std::io::Error>,
-        ));
-    }
-    if nonce.len() != 12 {
-        return Err(DsmError::crypto(
-            "Invalid nonce size for AES-GCM".to_string(),
-            None::<std::io::Error>,
-        ));
-    }
-
-    // Build cipher without touching GenericArray::from_slice
-    let cipher = Aes256Gcm::new_from_slice(key).map_err(|_| {
-        DsmError::crypto(
-            "Invalid AES-256 key length".to_string(),
-            None::<std::io::Error>,
-        )
-    })?;
-
-    let mut nonce_arr = [0u8; 12];
-    nonce_arr.copy_from_slice(nonce);
-    let nonce = Nonce::from(nonce_arr); // From<[u8;12]>, no removed APIs
-
-    cipher
-        .decrypt(&nonce, ciphertext)
-        .map_err(|_| DsmError::crypto("AES decryption failed".to_string(), None::<std::io::Error>))
-}
-
 // ------------------ Encoding helpers ------------------
 
 fn encode_u32(buf: &mut Vec<u8>, v: u32) {
@@ -671,85 +609,22 @@ mod tests {
     }
 
     #[test]
-    fn test_aes_encryption_decryption() {
-        let msg = b"Test AES-GCM encryption";
-        let key = vec![0xAAu8; 32]; // 256-bit key
-        let nonce = vec![0x55u8; 12]; // 96-bit nonce
+    fn test_kyber_keypair_rejects_trailing_bytes() {
+        let keypair = generate_kyber_keypair().expect("Key generation should succeed");
+        let mut bytes = keypair.to_bytes();
+        bytes.extend_from_slice(&[0xAB, 0xCD]);
 
-        let encrypted = aes_encrypt(&key, &nonce, msg);
-        assert!(encrypted.is_ok(), "AES encryption should succeed");
-
-        let ciphertext = encrypted.unwrap();
-
-        let decrypted = aes_decrypt(&key, &nonce, &ciphertext);
-        assert!(decrypted.is_ok(), "AES decryption should succeed");
-
-        assert_eq!(
-            msg,
-            decrypted.unwrap().as_slice(),
-            "Decrypted should match original"
-        );
+        let result = KyberKeyPair::from_bytes(&bytes);
+        assert!(result.is_err(), "Should reject trailing bytes");
     }
 
     #[test]
-    fn test_aes_wrong_key() {
-        let msg = b"Secret";
-        let key1 = vec![1u8; 32];
-        let key2 = vec![2u8; 32];
-        let nonce = vec![0u8; 12];
-
-        let ciphertext = aes_encrypt(&key1, &nonce, msg).expect("Encryption should succeed");
-
-        let decrypted = aes_decrypt(&key2, &nonce, &ciphertext);
-        assert!(
-            decrypted.is_err(),
-            "AES decryption with wrong key should fail"
-        );
-    }
-
-    #[test]
-    fn test_aes_wrong_nonce() {
-        let msg = b"Secret";
-        let key = vec![1u8; 32];
-        let nonce1 = vec![1u8; 12];
-        let nonce2 = vec![2u8; 12];
-
-        let ciphertext = aes_encrypt(&key, &nonce1, msg).expect("Encryption should succeed");
-
-        let decrypted = aes_decrypt(&key, &nonce2, &ciphertext);
-        assert!(
-            decrypted.is_err(),
-            "AES decryption with wrong nonce should fail"
-        );
-    }
-
-    #[test]
-    fn test_aes_invalid_key_size() {
-        let msg = b"Test";
-        let bad_key = vec![0u8; 16]; // Wrong size (should be 32)
-        let nonce = vec![0u8; 12];
-
-        let result = aes_encrypt(&bad_key, &nonce, msg);
-        assert!(result.is_err(), "Should reject invalid key size");
-    }
-
-    #[test]
-    fn test_aes_corrupted_ciphertext() {
-        let msg = b"Test message";
-        let key = vec![1u8; 32];
-        let nonce = vec![0u8; 12];
-
-        let mut ciphertext = aes_encrypt(&key, &nonce, msg).expect("Encryption should succeed");
-
-        // Corrupt the ciphertext
-        if !ciphertext.is_empty() {
-            ciphertext[0] ^= 0xFF;
-        }
-
-        let result = aes_decrypt(&key, &nonce, &ciphertext);
+    fn test_kyber_entropy_minimum_32_bytes() {
+        let short_entropy = vec![7u8; 31];
+        let result = generate_kyber_keypair_from_entropy(&short_entropy, "test");
         assert!(
             result.is_err(),
-            "Decryption of corrupted ciphertext should fail"
+            "Should reject entropy shorter than 32 bytes"
         );
     }
 
@@ -863,29 +738,5 @@ mod tests {
         assert_eq!(shared_secret1, shared_secret2);
         assert_eq!(ciphertext1, ciphertext2);
         assert_eq!(decapsulated, shared_secret1);
-    }
-
-    #[test]
-    fn test_aes_empty_message() {
-        let key = vec![1u8; 32];
-        let nonce = vec![0u8; 12];
-        let msg = b"";
-
-        let encrypted = aes_encrypt(&key, &nonce, msg).expect("Should encrypt empty message");
-        let decrypted = aes_decrypt(&key, &nonce, &encrypted).expect("Should decrypt");
-
-        assert_eq!(msg, decrypted.as_slice(), "Empty message should roundtrip");
-    }
-
-    #[test]
-    fn test_aes_large_message() {
-        let key = vec![1u8; 32];
-        let nonce = vec![0u8; 12];
-        let msg = vec![0xAAu8; 10000]; // 10 KB
-
-        let encrypted = aes_encrypt(&key, &nonce, &msg).expect("Should encrypt large message");
-        let decrypted = aes_decrypt(&key, &nonce, &encrypted).expect("Should decrypt");
-
-        assert_eq!(msg, decrypted, "Large message should roundtrip");
     }
 }
